moves general handlers to have them stored
