+++ rs/ai/requested_scaling/handlers/general_handlers/astrolabe_handler.py
