+++ rs/ai/requested_scaling/handlers/general_handlers/battle_handler.py
