+++ rs/ai/requested_scaling/handlers/general_handlers/boss_relic_handler.py
