+++ rs/ai/requested_scaling/handlers/general_handlers/campfire_handler.py
