+++ rs/ai/requested_scaling/handlers/general_handlers/card_reward_handler.py
