+++ rs/ai/requested_scaling/handlers/general_handlers/combat_reward_handler.py
