+++ rs/ai/requested_scaling/handlers/general_handlers/event_handler.py
