+++ rs/ai/requested_scaling/handlers/general_handlers/shop_entrance_handler.py
