+++ rs/ai/requested_scaling/handlers/general_handlers/shop_purchase_handler.py
