+++ rs/ai/requested_scaling/handlers/general_handlers/smart_battle_handler.py
