+++ rs/ai/requested_scaling/handlers/general_handlers/smart_path_handler.py
