+++ rs/ai/requested_scaling/handlers/general_handlers/upgrade_handler.py
