+++ rs/ai/requested_scaling/handlers/synergy_handlers/exhaust_action/exhaust_handler.py
@@ -66,3 +66,32 @@ class ExhaustHandler(Handler):
         if presentation_mode:
             return [p_delay, 'confirm']
         return ['confirm']
+
+
+
+
+
+top_list = [0,1,2,3,4]
+bot_list = [0,1,2,3,4]
+
+choose index 0 -> 0
+
+bot_list = [1,2,3,4]
+
+choose index 0 -> 1
+
+bot_list = [1,2,3]
+
+choose index 0 -> 2
+
+top_list.index(bot_list[index])
+
+top_list.index(2)
+
+
+
+
+
+
+
+
